Se crea migración de datos y se llena la tabla en la BD

+++ WebAPI_MVC_ASR/Models/Feed.cs
@@ -11,10 +11,10 @@ namespace WebAPI_MVC_ASR.Models
         public DateTimeOffset Timestamp { get; set; }
         public DateTimeOffset TimestampUTC { get; set; }
         public string? Country { get; set; }
-        public double? ImbalanceFee { get; set; }
-        public double? HourlyImbalanceFee { get; set; }
-        public double? PeakLoadFee { get; set; }
-        public double? VolumeFee { get; set; }
+        public float? ImbalanceFee { get; set; }
+        public float? HourlyImbalanceFee { get; set; }
+        public float? PeakLoadFee { get; set; }
+        public float? VolumeFee { get; set; }
         public int WeeklyFee { get; set; }
     }
 }
